pipeline_smp/requirements.txt
numpy
pandas
psycopg2
requests
sqlalchemy
python-dotenv
psutil
pyyaml